# docs/swagger.yaml
basePath: /
definitions:
  handler.Tokens:
    properties:
      accessToken:
        type: string
      exp:
        type: integer
      refreshToken:
        type: string
    type: object
  repository.Attachment:
    properties:
      id:
        type: integer
      label:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
  repository.Day:
    properties:
      attachments:
        items:
          $ref: '#/definitions/repository.Attachment'
        type: array
      description:
        type: string
      id:
        type: integer
      isLongRead:
        type: boolean
      title:
        type: string
    type: object
  repository.Project:
    properties:
      description:
        type: string
      id:
        type: integer
      link:
        type: string
      preview:
        type: string
      title:
        type: string
    type: object
  repository.Setting:
    properties:
      id:
        type: integer
      month:
        type: integer
      showAllDays:
        type: boolean
    type: object
  repository.User:
    properties:
      code:
        type: string
      email:
        type: string
      id:
        type: integer
      refreshToken:
        type: string
      role:
        type: string
    type: object
  validators.GlobalHandlerResp:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:9000
info:
  contact: {}
  title: Advent Calendar API docs
  version: "1.0"
paths:
  /api/days:
    get:
      parameters:
      - description: Например Europe/Samara
        in: query
        name: timeZone
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.Day'
            type: array
      tags:
      - Days
    post:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        minLength: 5
        name: description
        required: true
        type: string
      - in: formData
        minLength: 5
        name: title
        required: true
        type: string
      - collectionFormat: csv
        description: ' '
        in: formData
        items:
          type: file
        name: attachments
        type: array
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
      tags:
      - Days
  /api/days/{id}:
    put:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      - collectionFormat: csv
        in: formData
        items:
          type: integer
        name: attachmentIds
        type: array
      - in: formData
        minLength: 5
        name: description
        type: string
      - in: formData
        name: isLongRead
        type: boolean
      - in: formData
        minLength: 5
        name: title
        type: string
      - collectionFormat: csv
        description: ' '
        in: formData
        items:
          type: file
        name: attachments
        type: array
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
      tags:
      - Days
  /api/days/{id}/views:
    post:
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      - description: ' '
        in: formData
        name: email
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
      tags:
      - Days
  /api/days/admin:
    get:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.Day'
            type: array
      tags:
      - Days
  /api/projects:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.Project'
            type: array
      tags:
      - Projects
    post:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        minLength: 5
        name: description
        required: true
        type: string
      - in: formData
        name: link
        required: true
        type: string
      - in: formData
        minLength: 5
        name: title
        required: true
        type: string
      - description: ' '
        in: formData
        name: preview
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
      tags:
      - Projects
  /api/projects/{id}:
    delete:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
      tags:
      - Projects
    put:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      - in: formData
        minLength: 5
        name: description
        required: true
        type: string
      - in: formData
        name: link
        required: true
        type: string
      - in: formData
        minLength: 5
        name: title
        required: true
        type: string
      - description: ' '
        in: formData
        name: preview
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
      tags:
      - Projects
  /api/settings:
    get:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/repository.Setting'
      tags:
      - Settings
    put:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        maximum: 12
        minimum: 1
        name: month
        required: true
        type: integer
      - in: formData
        name: showAllDays
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
      tags:
      - Settings
  /api/users/check:
    get:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
      tags:
      - Users
  /api/users/confirm:
    patch:
      parameters:
      - in: formData
        name: code
        required: true
        type: string
      - in: formData
        minLength: 5
        name: email
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Tokens'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
      tags:
      - Users
  /api/users/login:
    post:
      parameters:
      - in: formData
        minLength: 5
        name: email
        required: true
        type: string
      - in: formData
        name: password
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Tokens'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
      tags:
      - Users
  /api/users/refresh:
    patch:
      parameters:
      - description: RefreshToken
        in: header
        name: RefreshToken
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Tokens'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
      tags:
      - Users
  /api/users/subscribe:
    delete:
      parameters:
      - in: query
        minLength: 5
        name: email
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
      tags:
      - Users
    post:
      parameters:
      - in: formData
        minLength: 5
        name: email
        required: true
        type: string
      - in: formData
        name: isConfirm
        required: true
        type: boolean
      - in: formData
        minLength: 5
        name: nickname
        required: true
        type: string
      - in: formData
        minLength: 5
        name: tgUsername
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validators.GlobalHandlerResp'
      tags:
      - Users
swagger: "2.0"